+++ src/lib/abstract-ops/internal-methods.ts
@@ -2,4 +2,5 @@ export const AbortSteps = Symbol('[[AbortSteps]]');
 export const ErrorSteps = Symbol('[[ErrorSteps]]');
 export const CancelSteps = Symbol('[[CancelSteps]]');
 export const PullSteps = Symbol('[[PullSteps]]');
+export const CanPullSyncSteps = Symbol('[[CanPullSyncSteps]]');
 export const ReleaseSteps = Symbol('[[ReleaseSteps]]');